Correction des droirts des utilisateurs sur les formulaires, protection de l'administrateur dans l'affiche des utilisateurs dans la liste permettant leur gestion, préparation du bouton de suppression d'un utilisateur, reste à gérer la mise en forme CSS et la fonction de suppression de la Dal

+++ controleurs/EvenementControleur.php
@@ -2,18 +2,23 @@
 //récupération des événements et des alertes:
 $evenements = recupererEvenements();
 $alertes = recupererAlertes();
+$rights = array();
+$rights[0] = "1";
+$rights[1] = "2";
 
-//traitement des formulaires de suppression:
-for($i=0; $i<sizeof($evenements); $i++){
-    if(isset($_POST['event'.$evenements[$i]->getId()])){
-        supprimerEvenementParId($evenements[$i]->getId());
-        $evenements = recupererEvenements();
+if(isset($_SESSION['user_id']) && !empty($_SESSION['user_id']) && in_array($_SESSION['user_right'], $rights, true)){
+    //traitement des formulaires de suppression:
+    for($i=0; $i<sizeof($evenements); $i++){
+        if(isset($_POST['event'.$evenements[$i]->getId()])){
+            supprimerEvenementParId($evenements[$i]->getId());
+            $evenements = recupererEvenements();
+        }
     }
-}
-for($i=0; $i<sizeof($alertes); $i++){
-    if(isset($_POST['alerte'.$alertes[$i]->getId()])){
-        supprimerAlerteParId($alertes[$i]->getId());
-        $alertes = recupererAlertes();
+    for($i=0; $i<sizeof($alertes); $i++){
+        if(isset($_POST['alerte'.$alertes[$i]->getId()])){
+            supprimerAlerteParId($alertes[$i]->getId());
+            $alertes = recupererAlertes();
+        }
     }
 }
